extract hourly slice utils function
